layers-review: compile fix (updating layers logic)
Small updates on the code:
 - new layers composition
 - new services naming (for WorkflowServer and Configuration)
 - adding comments, making the updates a bit clearer

For now, everything compiles and runs properly
